requirements.txt
django==3.2.16
django-environ
mysqlclient
channels
pyopenssl==21.0.0
cryptography==36.0.1
django-db-views
djangorestframework == 3.11.0
coreapi
requests_toolbelt
py_eureka_client